improve react perf
